gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.1
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
redis==5.0.1
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==23.2.0
prometheus-client==0.19.0
python-multipart==0.0.6
pyyaml==6.0.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0

# Code quality
ruff==0.1.6
black==23.11.0
mypy==1.7.1
pre-commit==3.5.0

# Monitoring and logging
loguru==0.7.2
